initial solution for the refactor. python installer now used instead of embedded packages, and removed most versions from the python version enum. Utilise the installer to remove the environment, and allowed multiplt versions of python to be installed in the python toolkit folder

+++ Python_Engine/Compute/Remove.cs
@@ -21,7 +21,11 @@
  */
 
 using BH.oM.Base.Attributes;
+using BH.oM.Python.Enums;
+using System;
+using System.Collections.Generic;
 using System.ComponentModel;
+using System.Diagnostics;
 using System.IO;
 
 namespace BH.Engine.Python
@@ -77,19 +81,59 @@ namespace BH.Engine.Python
             }
         }
 
-        [Description("Completely remove the base BHoM Python environment.")]
-        public static void RemoveBaseEnvironment()
+        [PreviousVersion("7.3", "BH.Engine.Python.Compute.RemoveBaseEnvironment()")]
+        [Input("version", "The base python version to remove.")]
+        public static void RemoveBaseVersion(PythonVersion version = PythonVersion.v3_10_11)
         {
-            string basePath = Path.Combine(Query.DirectoryEnvironments(), Query.ToolkitName());
-            if (Directory.Exists(basePath))
+            string basePath = Path.Combine(Query.DirectoryBaseEnvironment(), version.ToString());
+            string baseInstaller = Path.Combine(basePath, "installer.exe");
+
+            // If the installer does not exist and the folder does - assume a bad/invalid install and just delete the entire folder.
+            if (!File.Exists(baseInstaller) && Directory.Exists(basePath))
+            {
                 Directory.Delete(basePath, true);
+                return;
+            }
+            else if (!Directory.Exists(basePath))
+                // If the base path doesn't exist there is nothing to remove.
+                return;
+
+            using (Process uninstall = new Process()
+            {
+                StartInfo = new ProcessStartInfo()
+                {
+                    FileName = baseInstaller,
+                    Arguments = "/uninstall /passive",
+                    RedirectStandardError = true,
+                    UseShellExecute = false,
+                }
+            })
+            {
+                uninstall.Start();
+                string stderr = uninstall.StandardError.ReadToEnd();
+                uninstall.WaitForExit();
+                if (uninstall.ExitCode != 0)
+                {
+                    BH.Engine.Base.Compute.RecordError($"Error uninstalling python: {stderr}");
+                    return;
+                }
+            }
+
+            // Finally remove base folder as the installer may have missed something.
+            Directory.Delete(basePath, true);
         }
 
         [Description("Completely remove all BHoM-related Python environments and kernels.")]
         public static void RemoveEverything()
         {
             RemoveAllVirtualEnvironments();
-            RemoveBaseEnvironment();
+
+            //remove all python versions installed in base directory.
+            foreach (PythonVersion e in Enum.GetValues(typeof(PythonVersion)))
+            {
+                if (e != PythonVersion.Undefined)
+                    RemoveBaseVersion(e);
+            }
         }
     }
 }
